Add an optional bytes::BufMut function to RedisWrite (#1618)
* Add an optional bytes::BufMut function to RedisWrite

This allows for efficiently writing generated data without needing
an extra buffer (like serializing some data as ProtoBuf directly
into the command buffer).

This function is gated behind the "bytes" feature, which is
automatically enabled with any of the async features.

A default implementation is provided, so that this is not a breaking
change for upstream.

* Add a function to RedisWriter to reserve space for arguments

* Add tests and documentation for Cmd::bufmut_for_next_arg